Move ScopedOptions to pants.option.scope.
