+++ tests/python/pants_test/engine/legacy/test_options_parsing.py
@@ -6,11 +6,10 @@ from __future__ import absolute_import, division, print_function, unicode_litera
 
 from builtins import str
 
-from pants.engine.legacy.options_parsing import ScopedOptions
 from pants.engine.selectors import Params
 from pants.init.options_initializer import BuildConfigInitializer
 from pants.option.options_bootstrapper import OptionsBootstrapper
-from pants.option.scope import GLOBAL_SCOPE, Scope
+from pants.option.scope import GLOBAL_SCOPE, Scope, ScopedOptions
 from pants_test.test_base import TestBase
